refactor : 머지 후 수정
