+++ src/main/java/team7/hrbank/domain/employee/EmployeeStatus.java
@@ -1,0 +1,5 @@
+package team7.hrbank.domain.employee;
+
+public enum EmployeeStatus {
+    ACTIVE, ON_LEAVE, RESIGNED
+}
