--- src/main/java/team7/hrbank/domain/employee/entity/EmployeeStatus.java
@@ -1,5 +1,0 @@
-package team7.hrbank.domain.employee.entity;
-
-public enum EmployeeStatus {
-    ACTIVE,ON_LEAVE, RESIGNED
-}
